+++ src/main/java/fr/prefecture/sidsic/dashboard_sidsic/entity/Mouvement.java
@@ -15,7 +15,7 @@ public class Mouvement {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer ID;
+    private Long ID;
     @Version
     private int version;
