+++ src/main/java/fr/prefecture/sidsic/dashboard_sidsic/repository/AchatRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AchatRepository extends JpaRepository<Achat, Integer> {}
+public interface AchatRepository extends JpaRepository<Achat, Long> {}
