+++ src/main/java/fr/prefecture/sidsic/dashboard_sidsic/repository/MouvementRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MouvementRepository extends JpaRepository<Mouvement, Integer> {}
+public interface MouvementRepository extends JpaRepository<Mouvement, Long> {}
